vectorization/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.1
dill==0.3.8
docutils==0.21.2
environs==9.5.0
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
gradio==4.44.0
gradio_client==1.3.0
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
keyring==25.4.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
milvus-lite==2.4.10
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
nh3==0.2.18
numpy==2.1.1
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pkginfo==1.10.0
protobuf==5.28.2
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pymilvus==2.4.7
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
readme_renderer==44.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.8.1
ruff==0.6.8
safetensors==0.4.5
semantic-version==2.10.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
timm==1.0.9
tokenizers==0.20.0
tomlkit==0.12.0
torch==2.4.1
torchvision==0.19.1
towhee==1.1.3
towhee.models==1.1.3
tqdm==4.66.5
transformers==4.45.1
twine==5.1.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.31.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
xxhash==3.5.0
yarl==1.13.1
zipp==3.20.2
